pyproject.toml
[project]
name = "fygurs-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "gradio>=5.38.1",
    "ollama>=0.5.1",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
